`capture_tag` method
This method is used to create tags with nested erb markup.  Like the
`tag` method, it creates other tags with a similar API.

Unlike the `tag` method, it takes a block that contains nested markup
and does not take content as an arg.  Capture tags should be invoked
with the `<% ... %>`.